+++ packages/observable-client/src/chainHead/chainHead.ts
@@ -7,6 +7,7 @@ import {
   StorageResult,
 } from "@polkadot-api/substrate-client"
 import {
+  MonoTypeOperatorFunction,
   Observable,
   ReplaySubject,
   Subject,
@@ -72,24 +73,33 @@ export const getChainHead$ = (chainHead: ChainHead) => {
   const { withRecovery, withRecoveryFn } = getWithRecovery()
 
   const blockUsage$ = new Subject<BlockUsageEvent>()
+  const holdBlock = (hash: string) => {
+    blockUsage$.next({ type: "blockUsage", value: { type: "hold", hash } })
+    return () => {
+      setTimeout(() => {
+        blockUsage$.next({
+          type: "blockUsage",
+          value: { type: "release", hash },
+        })
+      }, 0)
+    }
+  }
+
+  const usingBlock: <T>(blockHash: string) => MonoTypeOperatorFunction<T> =
+    (blockHash: string) => (base) =>
+      new Observable((observer) => {
+        const release = holdBlock(blockHash)
+        const subscription = base.subscribe(observer)
+        subscription.add(release)
+        return subscription
+      })
+
   const withRefcount =
     <A extends Array<any>, T>(
       fn: (hash: string, ...args: A) => Observable<T>,
     ): ((hash: string, ...args: A) => Observable<T>) =>
     (hash, ...args) =>
-      new Observable((observer) => {
-        blockUsage$.next({ type: "blockUsage", value: { type: "hold", hash } })
-        const subscription = fn(hash, ...args).subscribe(observer)
-        return () => {
-          setTimeout(() => {
-            blockUsage$.next({
-              type: "blockUsage",
-              value: { type: "release", hash },
-            })
-          }, 0)
-          subscription.unsubscribe()
-        }
-      })
+      fn(hash, ...args).pipe(usingBlock(hash))
 
   const withInMemory =
     <A extends Array<any>, T>(
@@ -191,10 +201,10 @@ export const getChainHead$ = (chainHead: ChainHead) => {
 
     cache.set(hash, hashCache)
 
-    const connector = new ReplaySubject<T>()
+    let connector: ReplaySubject<T>
     const result = stream.pipe(
       share({
-        connector: () => connector,
+        connector: () => (connector = new ReplaySubject()),
       }),
       tap({
         complete() {
@@ -258,6 +268,7 @@ export const getChainHead$ = (chainHead: ChainHead) => {
   const withOptionalHash$ = getWithOptionalhash$(
     finalized$.pipe(map((b) => b.hash)),
     best$.pipe(map((b) => b.hash)),
+    usingBlock,
   )
 
   const _body$ = withOptionalHash$(commonEnhancer(lazyFollower("body")))
@@ -267,14 +278,23 @@ export const getChainHead$ = (chainHead: ChainHead) => {
   const _storage$ = commonEnhancer(lazyFollower("storage"))
 
   const storage$ = withOptionalHash$(
-    <Type extends StorageItemInput["type"], T>(
+    <
+      Type extends StorageItemInput["type"],
+      M extends
+        | undefined
+        | ((data: StorageResult<Type>, ctx: RuntimeContext) => any),
+    >(
       hash: string,
       withCanonicalChain: boolean,
       type: Type,
       keyMapper: (ctx: RuntimeContext) => string,
       childTrie: string | null = null,
-      mapper?: (data: StorageResult<Type>, ctx: RuntimeContext) => T,
-    ): Observable<unknown extends T ? StorageResult<Type> : T> =>
+      mapper?: M,
+    ): Observable<
+      undefined extends M
+        ? StorageResult<Type>
+        : { raw: StorageResult<Type>; mapped: ReturnType<NonNullable<M>> }
+    > =>
       pinnedBlocks$.pipe(
         take(1),
         mergeMap(
@@ -292,29 +312,31 @@ export const getChainHead$ = (chainHead: ChainHead) => {
             ? upsertCachedStream(
                 hash,
                 `storage-${type}-${key}-${childTrie ?? ""}-dec`,
-                unMapped$.pipe(map((x) => mapper(x, ctx))),
+                unMapped$.pipe(
+                  map((raw) => ({ raw, mapped: mapper(raw, ctx) })),
+                ),
               )
             : unMapped$
         }),
-      ) as Observable<unknown extends T ? StorageResult<Type> : T>,
+      ) as Observable<
+        undefined extends M
+          ? StorageResult<Type>
+          : { raw: StorageResult<Type>; mapped: ReturnType<NonNullable<M>> }
+      >,
   )
 
   const recoveralStorage$ = getRecoveralStorage$(getFollower, withRecovery)
   const storageQueries$ = withOptionalHash$(
-    withRefcount(
-      withStopRecovery(
-        pinnedBlocks$,
-        (hash: string, queries: Array<StorageItemInput>, childTrie?: string) =>
-          recoveralStorage$(hash, queries, childTrie ?? null, false),
-      ),
+    withStopRecovery(
+      pinnedBlocks$,
+      (hash: string, queries: Array<StorageItemInput>, childTrie?: string) =>
+        recoveralStorage$(hash, queries, childTrie ?? null, false),
     ),
   )
 
   const header$ = withOptionalHash$(
-    withRefcount(
-      withStopRecovery(pinnedBlocks$, (hash: string) =>
-        defer(() => getHeader(hash)),
-      ),
+    withStopRecovery(pinnedBlocks$, (hash: string) =>
+      defer(() => getHeader(hash)),
     ),
   )
 
@@ -326,7 +348,7 @@ export const getChainHead$ = (chainHead: ChainHead) => {
       (ctx) => ctx.events.key,
       null,
       (x, ctx) => ctx.events.dec(x!),
-    )
+    ).pipe(map((x) => x.mapped))
 
   const __call$ = commonEnhancer(lazyFollower("call"))
   const call$ = withOptionalHash$(
@@ -387,6 +409,7 @@ export const getChainHead$ = (chainHead: ChainHead) => {
       storageQueries$,
       eventsAt$: withCanonicalChain(eventsAt$),
 
+      holdBlock,
       trackTx$,
       trackTxWithoutEvents$,
       validateTx$,