+++ packages/signers/meta-signers/src/index.ts
@@ -1,0 +1,3 @@
+export type { PolkadotSigner } from "@polkadot-api/polkadot-signer"
+export * from "./multisig-signer"
+export * from "./proxy-signer"