+++ packages/signers/meta-signers/src/wrapped-signer.ts
@@ -1,0 +1,5 @@
+import { PolkadotSigner } from "@polkadot-api/polkadot-signer"
+
+export interface WrappedSigner extends PolkadotSigner {
+  accountId: Uint8Array
+}