+++ packages/merkleize-metadata/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 1.1.11 - 2024-12-10
+
+### Fixed
+
+- Update dependencies
+
 ## 1.1.10 - 2024-11-22
 
 ### Fixed