+++ packages/metadata-builders/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.9.3 - 2024-12-10
+
+### Fixed
+
+- Update dependencies
+
 ## 0.9.2 - 2024-11-22
 
 ### Fixed