+++ packages/substrate-bindings/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.10.0 - 2024-12-10
+
+### Added
+
+- `getMultisigAccountId` to get the AccountId of a multisig
+
 ## 0.9.4 - 2024-11-22
 
 ### Fixed